api/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
docker==4.0.2
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pylint==2.3.1
requests==2.22.0
retry==0.9.2
rope==0.14.0
six==1.12.0
SQLAlchemy==1.3.6
typed-ast==1.4.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
